requirements.txt
requests
pandas
ta
python-telegram-bot
python-dotenv
matplotlib
mplfinance